Make TypesCollection conform to DynamicMemberLookup only on Linux
