+++ SourceryRuntime/Sources/TemplateContext.swift
@@ -348,96 +348,3 @@ public final class Types: NSObject, SourceryModel, Diffable, DynamicMemberLookup
         })
     }()
 }
-
-/// :nodoc:
-#if canImport(ObjectiveC)
-@objcMembers
-#endif
-public class TypesCollection: NSObject, AutoJSExport {
-
-    // sourcery:begin: skipJSExport
-    let all: [Type]
-    let types: [String: [Type]]
-    let validate: ((Type) throws -> Void)?
-    // sourcery:end
-
-    init(types: [Type], collection: (Type) -> [String], validate: ((Type) throws -> Void)? = nil) {
-        self.all = types
-        var content = [String: [Type]]()
-        self.all.forEach { type in
-            collection(type).forEach { name in
-                var list = content[name] ?? [Type]()
-                list.append(type)
-                content[name] = list
-            }
-        }
-        self.types = content
-        self.validate = validate
-    }
-
-    public func types(forKey key: String) throws -> [Type] {
-        // In some configurations, the types are keyed by "ModuleName.TypeName"
-        var longKey: String?
-
-        if let validate = validate {
-            guard let type = all.first(where: { $0.name == key }) else {
-                throw "Unknown type \(key), should be used with `based`"
-            }
-
-            try validate(type)
-
-            if let module = type.module {
-                longKey = [module, type.name].joined(separator: ".")
-            }
-        }
-
-        // If we find the types directly, return them
-        if let types = types[key] {
-            return types
-        }
-
-        // if we find a types for the longKey, return them
-        if let longKey = longKey, let types = types[longKey] {
-            return types
-        }
-
-        return []
-    }
-
-    /// :nodoc:
-#if canImport(ObjectiveC)
-    override public func value(forKey key: String) -> Any? {
-        do {
-            return try types(forKey: key)
-        } catch {
-            Log.error(error)
-            return nil
-        }
-    }
-#else
-public func value(forKey key: String) -> Any? {
-        do {
-            return try types(forKey: key)
-        } catch {
-            Log.error(error)
-            return nil
-        }
-    }
-#endif
-
-    /// :nodoc:
-    public subscript(_ key: String) -> [Type] {
-        do {
-            return try types(forKey: key)
-        } catch {
-            Log.error(error)
-            return []
-        }
-    }
-
-#if canImport(ObjectiveC)
-    override public func responds(to aSelector: Selector!) -> Bool {
-        return true
-    }
-#endif
-}
