+++ SourceryRuntime/Sources/TypesCollection_Linux.swift
@@ -1,0 +1,88 @@
+//
+// Created by Krzysztof Zablocki on 31/12/2016.
+// Copyright (c) 2016 Pixle. All rights reserved.
+//
+
+import Foundation
+// For DynamicMemberLookup we need to import Stencil,
+// however, this is different from SourceryRuntime.content.generated.swift, because
+// it cannot reference Stencil
+import Stencil
+
+/// :nodoc:
+#if !canImport(ObjectiveC)
+public class TypesCollection: NSObject, AutoJSExport, DynamicMemberLookup {
+    public subscript(dynamicMember member: String) -> Any? {
+        return try? types(forKey: member)
+    }
+
+    // sourcery:begin: skipJSExport
+    let all: [Type]
+    let types: [String: [Type]]
+    let validate: ((Type) throws -> Void)?
+    // sourcery:end
+
+    init(types: [Type], collection: (Type) -> [String], validate: ((Type) throws -> Void)? = nil) {
+        self.all = types
+        var content = [String: [Type]]()
+        self.all.forEach { type in
+            collection(type).forEach { name in
+                var list = content[name] ?? [Type]()
+                list.append(type)
+                content[name] = list
+            }
+        }
+        self.types = content
+        self.validate = validate
+    }
+
+    public func types(forKey key: String) throws -> [Type] {
+        // In some configurations, the types are keyed by "ModuleName.TypeName"
+        var longKey: String?
+
+        if let validate = validate {
+            guard let type = all.first(where: { $0.name == key }) else {
+                throw "Unknown type \(key), should be used with `based`"
+            }
+
+            try validate(type)
+
+            if let module = type.module {
+                longKey = [module, type.name].joined(separator: ".")
+            }
+        }
+
+        // If we find the types directly, return them
+        if let types = types[key] {
+            return types
+        }
+
+        // if we find a types for the longKey, return them
+        if let longKey = longKey, let types = types[longKey] {
+            return types
+        }
+
+        return []
+    }
+
+    /// :nodoc:
+    public func value(forKey key: String) -> Any? {
+        do {
+            return try types(forKey: key)
+        } catch {
+            Log.error(error)
+            return nil
+        }
+    }
+
+    /// :nodoc:
+    public subscript(_ key: String) -> [Type] {
+        do {
+            return try types(forKey: key)
+        } catch {
+            Log.error(error)
+            return []
+        }
+    }
+}
+#endif
